requirements.txt
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
connexion==2.14.2
exceptiongroup==1.1.0
flake8==2.5.5
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mysql-connector-python==8.0.32
packaging==23.0
pep8==1.7.1
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.10.0
pyflakes==1.0.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.47
str2bool==1.1
urllib3==1.26.14
bleach==6.0.0
loguru==0.6.0
pdfkit==0.6.1
pytest-order==1.1.0
pytest-cov==4.0.0
xlsxwriter==3.1.0
pyhtml2pdf==0.0.6
pyvista==0.39.1
numpy==1.24.2
pymeshfix==0.16.2